requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
PyMySQL==1.1.0
reportlab==4.0.7
cryptography==41.0.7